Remove `DesignTimeHost` dependency.
- This resulted in us decoupling ourselves from the `IPlugin` interface. It's going away.
- Redesigned the `TagHelperDescriptor` resolver to be a glorified factory with a protocol. It used to serialize/deserialize messages as input because it was an `IPlugin`.

#25
