--- src/Microsoft.AspNet.Tooling.Razor/Models/OutgoingMessages/ResolveTagHelperDescriptorsMessage.cs
@@ -1,13 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-namespace Microsoft.AspNet.Tooling.Razor.Models.OutgoingMessages
-{
-    public class ResolveTagHelperDescriptorsMessage : RazorPluginMessage<ResolveTagHelperDescriptorsResponseData>
-    {
-        public ResolveTagHelperDescriptorsMessage(ResolveTagHelperDescriptorsResponseData data)
-            : base(messageType: RazorPluginMessageTypes.ResolveTagHelperDescriptors, data: data)
-        {
-        }
-    }
-}
