--- src/Microsoft.AspNet.Tooling.Razor/RazorPluginMessageTypes.cs
@@ -1,10 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-namespace Microsoft.AspNet.Tooling.Razor
-{
-    public static class RazorPluginMessageTypes
-    {
-        public const string ResolveTagHelperDescriptors = nameof(ResolveTagHelperDescriptors);
-    }
-}
