+++ src/Microsoft.AspNet.Tooling.Razor/project.json
@@ -6,10 +6,6 @@
   },
   "dependencies": {
     "Microsoft.AspNet.Razor.Runtime": "4.0.0-*",
-    "Microsoft.Dnx.DesignTimeHost.Abstractions": {
-      "version": "1.0.0-*",
-      "type": "build"
-    },
     "Microsoft.Extensions.PlatformAbstractions": {
       "version": "1.0.0-*",
       "type": "build"
