+++ test/Microsoft.AspNet.Tooling.Razor.Test/ResolveProtocolCommandTest.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using Microsoft.AspNet.Tooling.Razor.Internal;
 using Xunit;
 
 namespace Microsoft.AspNet.Tooling.Razor
