+++ test/Microsoft.AspNet.Tooling.Razor.Test/project.json
@@ -16,8 +16,5 @@
   },
   "commands": {
     "test": "xunit.runner.aspnet"
-  },
-  "compilationOptions": {
-    "keyFile": "../../tools/Key.snk"
   }
 }
